Forcing

VortexPasta.Forcing — Module

Forcing

Defines methods for injecting energy onto a system of vortices.

Normal fluid forcing

VortexPasta.Forcing.NormalFluidForcing — Type

NormalFluidForcing <: AbstractForcing
NormalFluidForcing(vn::Function; α, α′ = 0)

Forcing due to mutual friction with a normal fluid.

The normal fluid is represented by a function vn which should take a position x⃗::SVector{N, T} and return a velocity v⃗::SVector{N, T} (N is the number of dimensions, usually N = 3).

In particular, the function could be a synthetic velocity field from the SyntheticFields module (see below for examples).

This type of forcing defines an external velocity $\bm{v}_{\text{f}}$ affecting vortex motion, so that the actual vortex velocity $\bm{v}_{\text{L}}$ is

\[\frac{\mathrm{d}\bm{s}}{\mathrm{d}t} = \bm{v}_{\text{L}} = \bm{v}_{\text{s}} + \bm{v}_{\text{f}}.\]

Here $\bm{v}_{\text{s}}$ is the self-induced velocity obtained by applying Biot–Savart's law.

The forcing velocity is of the form:

\[\bm{v}_{\text{f}} = α \bm{s}' × (\bm{v}_{\text{n}} - \bm{v}_{\text{s}}) - α' \bm{s}' × \left[ \bm{s}' × (\bm{v}_{\text{n}} - \bm{v}_{\text{s}}) \right]\]

where $\bm{s}'$ is the local unit tangent vector, and $α$ and $α'$ are non-dimensional coefficients representing the intensity of Magnus and drag forces.

Example

Define a mutual friction forcing based on a large-scale normal fluid velocity field (see SyntheticFields.FourierBandVectorField):

julia> using VortexPasta.Forcing: NormalFluidForcing

julia> using VortexPasta.SyntheticFields: SyntheticFields, FourierBandVectorField

julia> using Random: Xoshiro

julia> rng = Xoshiro(42);  # initialise random number generator (optional, but recommended)

julia> Ls = (2π, 2π, 2π);  # domain dimensions

julia> vn_rms = 1.0;  # typical magnitude (rms value) of normal fluid velocity components

julia> vn = FourierBandVectorField(undef, Ls; kmin = 0.1, kmax = 1.5)  # create field with non-zero Fourier wavevectors kmin ≤ |k⃗| ≤ kmax
FourierBandVectorField{Float64, 3} with 9 independent Fourier coefficients in |k⃗| ∈ [1.0, 1.4142]

julia> SyntheticFields.init_coefficients!(rng, vn, vn_rms);  # randomly set non-zero Fourier coefficients of the velocity field

julia> forcing = NormalFluidForcing(vn; α = 0.8, α′ = 0)
NormalFluidForcing{Float64} with:
 ├─ Magnus force coefficient: α = 0.8
 ├─ Drag force coefficient: α′ = 0.0
 └─ Normal velocity field: FourierBandVectorField{Float64, 3} with 9 independent Fourier coefficients in |k⃗| ∈ [1.0, 1.4142]

VortexPasta.Forcing.apply! — Function

Forcing.apply!(forcing::AbstractForcing, vs::AbstractVector{<:Vec3}, f::AbstractFilament)

Apply forcing to a single filament f with self-induced velocities vs.

At output, the vs vector is overwritten with the actual vortex line velocities.

Forcing.apply!(forcing::NormalFluidForcing, vs, vn, tangents)

This variant can be used in the case of a NormalFluidForcing if one already has precomputed values of the normal fluid velocity and local unit tangents at filament points.

The normal fluid velocities may be computed using Forcing.get_velocities!.

VortexPasta.Forcing.get_velocities! — Function

Forcing.get_velocities!(forcing::NormalFluidForcing, vn::AbstractVector{<:Vec3}, f::AbstractFilament)

Evaluate normal fluid velocity at filament nodes.

Results are written to vn, which should have the same length as the filament f.

Abstract types

VortexPasta.Forcing.AbstractForcing — Type

AbstractForcing

Abstract type representing a forcing method.